requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
webdriver-manager
requests
json
faker
Appium-Python-Client

